Allows update of old pm token after chain of fails
So if I have a payment method to update, and there's an error on the first attempt, it won't get saved to the DB. Thus, the payment token to update must propagate through the attempts. Before, it would point to the previous failed attempt, which wasn't in the DB of course.